feat: refactor routes use blueprint to notification-template
