--- apps/api/src/app/blueprint/usecases/create-blueprint/create-blueprint.command.ts
@@ -1,8 +1,0 @@
-import { IsDefined, IsMongoId } from 'class-validator';
-import { EnvironmentWithUserCommand } from '../../../shared/commands/project.command';
-
-export class CreateBlueprintCommand extends EnvironmentWithUserCommand {
-  @IsDefined()
-  @IsMongoId()
-  templateId: string;
-}
