--- apps/api/src/app/blueprint/usecases/create-blueprint/index.ts
@@ -1,2 +1,0 @@
-export * from './create-blueprint.usecase';
-export * from './create-blueprint.command';
