+++ apps/api/src/app/blueprint/usecases/index.ts
@@ -1,10 +1,8 @@
-import { CreateBlueprint } from './create-blueprint';
 import { GetBlueprint } from './get-blueprint';
-import { GetGroupedBlueprints } from './get-blueprints';
+import { GetGroupedBlueprints } from './get-grouped-blueprints';
 
 export const USE_CASES = [
   //
-  CreateBlueprint,
   GetBlueprint,
   GetGroupedBlueprints,
 ];
