+++ apps/api/src/app/notification-template/dto/create-notification-template.request.dto.ts
@@ -14,7 +14,7 @@ export class CreateNotificationTemplateRequestDto implements ICreateNotification
   @ApiProperty()
   @IsString()
   @IsDefined({
-    message: 'Notification group must be provided',
+    message: 'Notification group must be provided ',
   })
   notificationGroupId: string;
 
@@ -57,4 +57,8 @@ export class CreateNotificationTemplateRequestDto implements ICreateNotification
   })
   @IsOptional()
   preferenceSettings?: IPreferenceChannels;
+
+  @ApiPropertyOptional({})
+  @IsOptional()
+  blueprintId?: string;
 }
