+++ apps/api/src/app/notification-template/e2e/change-template-status.e2e.ts
@@ -23,14 +23,14 @@ describe('Change template status by id - /notification-templates/:templateId/sta
     });
     const beforeChange = await notificationTemplateRepository.findById(template._id, template._environmentId);
 
-    expect(beforeChange.active).to.equal(false);
-    expect(beforeChange.draft).to.equal(true);
+    expect(beforeChange?.active).to.equal(false);
+    expect(beforeChange?.draft).to.equal(true);
     const { body } = await session.testAgent.put(`/v1/notification-templates/${template._id}/status`).send({
       active: true,
     });
     const found = await notificationTemplateRepository.findById(template._id, template._environmentId);
 
-    expect(found.active).to.equal(true);
-    expect(found.draft).to.equal(false);
+    expect(found?.active).to.equal(true);
+    expect(found?.draft).to.equal(false);
   });
 });
