+++ apps/api/src/app/notification-template/notification-template.controller.ts
@@ -170,6 +170,7 @@ export class NotificationTemplateController {
         draft: body.draft ?? true,
         critical: body.critical ?? false,
         preferenceSettings: body.preferenceSettings,
+        blueprintId: body.blueprintId,
       })
     );
   }
