+++ apps/api/src/app/notification-template/usecases/create-notification-template/create-notification-template.usecase.ts
@@ -2,16 +2,20 @@ import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import slugify from 'slugify';
 import * as shortid from 'shortid';
 
-import { NotificationTemplateRepository } from '@novu/dal';
+import {
+  NotificationTemplateRepository,
+  NotificationGroupEntity,
+  NotificationStepEntity,
+  FeedRepository,
+  NotificationGroupRepository,
+} from '@novu/dal';
 import { ChangeEntityTypeEnum, INotificationTemplateStep, INotificationTrigger, TriggerTypeEnum } from '@novu/shared';
 import { AnalyticsService } from '@novu/application-generic';
 
 import { CreateNotificationTemplateCommand } from './create-notification-template.command';
 import { ContentService } from '../../../shared/helpers/content.service';
-import { CreateMessageTemplate } from '../../../message-template/usecases/create-message-template/create-message-template.usecase';
-import { CreateMessageTemplateCommand } from '../../../message-template/usecases/create-message-template/create-message-template.command';
+import { CreateMessageTemplate, CreateMessageTemplateCommand } from '../../../message-template/usecases';
 import { CreateChange, CreateChangeCommand } from '../../../change/usecases';
-
 import { ApiException } from '../../../shared/exceptions/api.exception';
 
 @Injectable()
@@ -19,11 +23,17 @@ export class CreateNotificationTemplate {
   constructor(
     private notificationTemplateRepository: NotificationTemplateRepository,
     private createMessageTemplate: CreateMessageTemplate,
+    private notificationGroupRepository: NotificationGroupRepository,
+    private feedRepository: FeedRepository,
     private createChange: CreateChange,
     private analyticsService: AnalyticsService
   ) {}
 
-  async execute(command: CreateNotificationTemplateCommand) {
+  async execute(usecaseCommand: CreateNotificationTemplateCommand) {
+    const blueprintCommand = await this.processBlueprint(usecaseCommand);
+
+    const command = blueprintCommand ?? usecaseCommand;
+
     const contentService = new ContentService();
     const variables = contentService.extractMessageVariables(command.steps);
     const subscriberVariables = contentService.extractSubscriberMessageVariables(command.steps);
@@ -138,6 +148,112 @@ export class CreateNotificationTemplate {
       });
     }
 
+    if (command.blueprintId) {
+      await this.analyticsService.track('[Notification directory] - Template created from blueprint', command.userId, {
+        blueprintId: command.blueprintId,
+        environmentId: command.environmentId,
+        organizationId: command.organizationId,
+      });
+    }
+
     return item;
+  }
+
+  private async processBlueprint(command: CreateNotificationTemplateCommand) {
+    if (!command.blueprintId) return null;
+
+    const group: NotificationGroupEntity = await this.handleGroup(command.notificationGroupId, command);
+    const steps: NotificationStepEntity[] = await this.handleFeeds(command.steps as any, command);
+
+    return CreateNotificationTemplateCommand.create({
+      organizationId: command.organizationId,
+      userId: command.userId,
+      environmentId: command.environmentId,
+      name: command.name,
+      tags: command.tags,
+      description: command.description,
+      steps: steps,
+      notificationGroupId: group._id,
+      active: command.active ?? false,
+      draft: command.draft ?? true,
+      critical: command.critical ?? false,
+      preferenceSettings: command.preferenceSettings,
+      blueprintId: command.blueprintId,
+    });
+  }
+
+  private async handleFeeds(
+    steps: NotificationStepEntity[],
+    command: CreateNotificationTemplateCommand
+  ): Promise<NotificationStepEntity[]> {
+    for (let i = 0; i < steps.length; i++) {
+      const step = steps[i];
+
+      if (!step.template?._feedId) {
+        continue;
+      }
+
+      const blueprintFeed = await this.feedRepository.findOne({
+        _organizationId: this.getBlueprintOrganizationId,
+        id: step.template._feedId,
+      });
+
+      if (!blueprintFeed) {
+        step.template._feedId = undefined;
+        steps[i] = step;
+        continue;
+      }
+
+      let foundFeed = await this.feedRepository.findOne({
+        _organizationId: command.organizationId,
+        identifier: blueprintFeed.identifier,
+      });
+
+      if (!foundFeed) {
+        foundFeed = await this.feedRepository.create({
+          name: blueprintFeed.name,
+          identifier: blueprintFeed.identifier,
+          _environmentId: command.environmentId,
+          _organizationId: command.organizationId,
+        });
+      }
+
+      step.template._feedId = foundFeed._id;
+      steps[i] = step;
+    }
+
+    return steps;
+  }
+
+  private async handleGroup(
+    notificationGroupId: string,
+    command: CreateNotificationTemplateCommand
+  ): Promise<NotificationGroupEntity> {
+    const blueprintNotificationGroup = await this.notificationGroupRepository.findOne({
+      _id: notificationGroupId,
+      _organizationId: this.getBlueprintOrganizationId,
+    });
+
+    if (!blueprintNotificationGroup)
+      throw new NotFoundException(`Blueprint notification group with id ${notificationGroupId} is not found`);
+
+    let group = await this.notificationGroupRepository.findOne({
+      name: blueprintNotificationGroup.name,
+      _environmentId: command.environmentId,
+      _organizationId: command.organizationId,
+    });
+
+    if (!group) {
+      group = await this.notificationGroupRepository.create({
+        _environmentId: command.environmentId,
+        _organizationId: command.organizationId,
+        name: blueprintNotificationGroup.name,
+      });
+    }
+
+    return group;
+  }
+  private get getBlueprintOrganizationId(): string {
+    return NotificationTemplateRepository.getBlueprintOrganizationId() as string;
   }
 }
