+++ apps/web/src/pages/templates/components/BlueprintModal.tsx
@@ -5,8 +5,8 @@ import { useQuery, useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 import { updateUserOnBoarding } from '../../../api/user';
-import { IUserEntity } from '@novu/shared';
-import { createTemplateFromBluePrintId, getBlueprintTemplateById } from '../../../api/notification-templates';
+import { ICreateNotificationTemplateDto, IUserEntity } from '@novu/shared';
+import { createTemplate, getBlueprintTemplateById } from '../../../api/notification-templates';
 import { errorMessage } from '../../../utils/notifications';
 import { When } from '../../../components/utils/When';
 import { useSegment } from '../../../components/providers/SegmentProvider';
@@ -51,7 +51,7 @@ export function BlueprintModal() {
     }
   );
 
-  const { mutate, isLoading: isCreating } = useMutation(createTemplateFromBluePrintId, {
+  const { mutate, isLoading: isCreating } = useMutation(createTemplate, {
     onSuccess: (template) => {
       if (template) {
         disableOnboarding();
@@ -70,6 +70,14 @@ export function BlueprintModal() {
   });
 
   const isLoading = isBluePrintLoading || isCreating;
+
+  function mapBlueprintToTemplateCreatePayload(): ICreateNotificationTemplateDto {
+    const templatePayload = Object.assign({}, blueprint);
+    templatePayload.notificationGroupId = templatePayload._notificationGroupId;
+    templatePayload.blueprintId = templatePayload._id;
+
+    return templatePayload;
+  }
 
   return (
     <>
@@ -118,7 +126,8 @@ export function BlueprintModal() {
             data-test-id="create-from-blueprint"
             onClick={() => {
               if (blueprintId) {
-                mutate(blueprintId);
+                const templatePayload = mapBlueprintToTemplateCreatePayload();
+                mutate(templatePayload);
               }
             }}
           >
